+++ samples/Store/Integration/FavoritesIntegration.fs
@@ -1,5 +1,6 @@
 ﻿module Samples.Store.Integration.FavoritesIntegration
 
+open Domain
 open Equinox
 open Equinox.CosmosStore.Integration
 open Swensen.Unquote
@@ -11,27 +12,27 @@ let snapshot = Domain.Favorites.Fold.isOrigin, Domain.Favorites.Fold.snapshot
 
 let createMemoryStore () = MemoryStore.VolatileStore<_>()
 let createServiceMemory log store =
-    Backend.Favorites.create log (MemoryStore.Resolver(store, FsCodec.Box.Codec.Create(), fold, initial).Resolve)
+    Favorites.create log (MemoryStore.Resolver(store, FsCodec.Box.Codec.Create(), fold, initial).Resolve)
 
 let codec = Domain.Favorites.Events.codec
 let createServiceGes log gateway =
     let resolver = EventStore.Resolver(gateway, codec, fold, initial, access = EventStore.AccessStrategy.RollingSnapshots snapshot)
-    Backend.Favorites.create log resolver.Resolve
+    Favorites.create log resolver.Resolve
 
 let createServiceCosmos log context =
     let category = CosmosStore.CosmosStoreCategory(context, codec, fold, initial, CosmosStore.CachingStrategy.NoCaching, CosmosStore.AccessStrategy.Snapshot snapshot)
-    Backend.Favorites.create log category.Resolve
+    Favorites.create log category.Resolve
 
 let createServiceCosmosRollingState log context =
     let access = CosmosStore.AccessStrategy.RollingState Domain.Favorites.Fold.snapshot
     let category = CosmosStore.CosmosStoreCategory(context, codec, fold, initial, CosmosStore.CachingStrategy.NoCaching, access)
-    Backend.Favorites.create log category.Resolve
+    Favorites.create log category.Resolve
 
 type Tests(testOutputHelper) =
     let testOutput = TestOutputAdapter testOutputHelper
     let createLog () = createLogger testOutput
 
-    let act (service : Backend.Favorites.Service) (clientId, command) = async {
+    let act (service : Favorites.Service) (clientId, command) = async {
         do! service.Execute(clientId, command)
         let! items = service.List clientId